Migrated Back Button feature logic and implementation to viewmodel and main_view
