+++ lib/ui/views/settings/settings_view.dart
@@ -11,135 +11,113 @@ import 'package:stacked/stacked.dart';
 // in this case, is not a "Page"
 // it's just a UI representation.
 
-// ignore: must_be_immutable
 class SettingsView extends StatelessWidget {
-  SettingsView({
+  const SettingsView({
     Key? key,
   }) : super(key: key);
-
-  static const snackBarDuration = Duration(seconds: 2);
-
-  final snackBar = const SnackBar(
-    content: Text(
-      "Press Back again to quit the app",
-      textAlign: TextAlign.center,
-    ),
-    duration: snackBarDuration,
-  );
-
-  DateTime backButtonPressTime = DateTime.utc(1989, DateTime.november, 9);
 
   @override
   Widget build(BuildContext context) {
     final customTheme = Theme.of(context).extension<CustomTheme>()!;
 
-    return Scaffold(
-      body: Builder(
-        builder: (context) {
-          return WillPopScope(
-            onWillPop: () => handleWillPop(context),
-            child: ViewModelBuilder<SettingsViewModel>.reactive(
-                viewModelBuilder: () => SettingsViewModel(),
-                builder: (context, model, child) {
-                  return SingleChildScrollView(
-                    physics: const BouncingScrollPhysics(),
-                    child: Padding(
-                        padding: const EdgeInsets.fromLTRB(6, 6, 6, 15),
-                        child: Column(
-                          children: [
-                            _buildCard(
-                              children: [
-                                ListTile(
-                                  contentPadding: const EdgeInsets.fromLTRB(15, 15, 0, 15),
-                                  leading: CircleAvatar(
-                                    radius: 22,
-                                    backgroundColor: customTheme.avatarBackgroundColor,
-                                    child: Icon(
-                                      Icons.person_rounded,
-                                      color: customTheme.settingsIconColor,
-                                    ),
-                                  ),
-                                  title: const ThemeText.heading4(
-                                    "Jonel Dominic Tapang",
-                                  ),
-                                  trailing: IconButton(
-                                    iconSize: 28,
-                                    color: customTheme.actionButtonColor,
-                                    onPressed: () {},
-                                    icon: const Icon(Icons.chevron_right_rounded),
-                                  ),
-                                ),
-                              ],
+    return ViewModelBuilder<SettingsViewModel>.reactive(
+        viewModelBuilder: () => SettingsViewModel(),
+        builder: (context, model, child) {
+          return SingleChildScrollView(
+            physics: const BouncingScrollPhysics(),
+            child: Padding(
+                padding: const EdgeInsets.fromLTRB(6, 6, 6, 15),
+                child: Column(
+                  children: [
+                    _buildCard(
+                      children: [
+                        ListTile(
+                          contentPadding: const EdgeInsets.fromLTRB(15, 15, 0, 15),
+                          leading: CircleAvatar(
+                            radius: 22,
+                            backgroundColor: customTheme.avatarBackgroundColor,
+                            child: Icon(
+                              Icons.person_rounded,
+                              color: customTheme.settingsIconColor,
                             ),
-                            verticalSpaceVeryTiny,
-                            _buildCard(
-                              children: [
-                                _buildListTile(
-                                  iconData: Icons.account_balance_rounded,
-                                  title: 'Accounts',
-                                  onPressed: model.navigateToAccountSettings,
-                                  theme: customTheme,
-                                ),
-                                _buildListTile(
-                                  iconData: Icons.category_rounded,
-                                  title: 'Categories',
-                                  onPressed: model.navigateToCategoryList,
-                                  theme: customTheme,
-                                ),
-                                _buildListTile(
-                                  iconData: Icons.notification_important_rounded,
-                                  title: 'Notifications',
-                                  onPressed: () {},
-                                  theme: customTheme,
-                                ),
-                              ],
+                          ),
+                          title: const ThemeText.heading4(
+                            "Jonel Dominic Tapang",
+                          ),
+                          trailing: IconButton(
+                            iconSize: 28,
+                            color: customTheme.actionButtonColor,
+                            onPressed: () {},
+                            icon: const Icon(Icons.chevron_right_rounded),
+                          ),
+                        ),
+                      ],
+                    ),
+                    verticalSpaceVeryTiny,
+                    _buildCard(
+                      children: [
+                        _buildListTile(
+                          iconData: Icons.account_balance_rounded,
+                          title: 'Accounts',
+                          onPressed: model.navigateToAccountSettings,
+                          theme: customTheme,
+                        ),
+                        _buildListTile(
+                          iconData: Icons.category_rounded,
+                          title: 'Categories',
+                          onPressed: model.navigateToCategoryList,
+                          theme: customTheme,
+                        ),
+                        _buildListTile(
+                          iconData: Icons.notification_important_rounded,
+                          title: 'Notifications',
+                          onPressed: () {},
+                          theme: customTheme,
+                        ),
+                      ],
+                    ),
+                    verticalSpaceVeryTiny,
+                    _buildCard(
+                      children: [
+                        SwitchListTile(
+                          title: const ThemeText.heading4(
+                            'Dark Mode',
+                          ),
+                          secondary: SizedBox(
+                            width: 44,
+                            child: Icon(
+                              Icons.dark_mode_rounded,
+                              color: customTheme.settingsIconColor,
                             ),
-                            verticalSpaceVeryTiny,
-                            _buildCard(
-                              children: [
-                                SwitchListTile(
-                                  title: const ThemeText.heading4(
-                                    'Dark Mode',
-                                  ),
-                                  secondary: SizedBox(
-                                    width: 44,
-                                    child: Icon(
-                                      Icons.dark_mode_rounded,
-                                      color: customTheme.settingsIconColor,
-                                    ),
-                                  ),
-                                  value: model.getIsThemeModeIsDark(),
-                                  onChanged: model.toggleTheme,
-                                  activeColor: customTheme.activeControlColor,
-                                  activeTrackColor: customTheme.activeSwitchTrackColor,
-                                )
-                              ],
-                            ),
-                            verticalSpaceVeryTiny,
-                            _buildCard(
-                              children: [
-                                _buildListTile(
-                                  iconData: Icons.rate_review_rounded,
-                                  title: 'Feedback',
-                                  onPressed: () {},
-                                  theme: customTheme,
-                                ),
-                                _buildListTile(
-                                  iconData: Icons.info_rounded,
-                                  title: 'About',
-                                  onPressed: () {},
-                                  theme: customTheme,
-                                ),
-                              ],
-                            ),
-                          ],
-                        )),
-                  );
-                }),
+                          ),
+                          value: model.getIsThemeModeIsDark(),
+                          onChanged: model.toggleTheme,
+                          activeColor: customTheme.activeControlColor,
+                          activeTrackColor: customTheme.activeSwitchTrackColor,
+                        )
+                      ],
+                    ),
+                    verticalSpaceVeryTiny,
+                    _buildCard(
+                      children: [
+                        _buildListTile(
+                          iconData: Icons.rate_review_rounded,
+                          title: 'Feedback',
+                          onPressed: () {},
+                          theme: customTheme,
+                        ),
+                        _buildListTile(
+                          iconData: Icons.info_rounded,
+                          title: 'About',
+                          onPressed: () {},
+                          theme: customTheme,
+                        ),
+                      ],
+                    ),
+                  ],
+                )),
           );
-        },
-      ),
-    );
+        });
   }
 
   ListTile _buildListTile({
@@ -173,21 +151,5 @@ class SettingsView extends StatelessWidget {
         children: children,
       ),
     );
-  }
-
-  final GlobalKey<ScaffoldMessengerState> snackbarKey = GlobalKey<ScaffoldMessengerState>();
-
-  Future<bool> handleWillPop(BuildContext context) async {
-    final now = DateTime.now();
-    final backButtonHasNotBeenPressedOrSnackBarHasBeenClosed =
-        now.difference(backButtonPressTime) > snackBarDuration;
-
-    if (backButtonHasNotBeenPressedOrSnackBarHasBeenClosed) {
-      backButtonPressTime = now;
-      ScaffoldMessenger.of(context).showSnackBar(snackBar);
-      return false;
-    }
-
-    return true;
   }
 }
